+++ utils/src/main/java/com/seewo/student/libutils/utils/PowerSaveModeHelper.kt
@@ -37,7 +37,7 @@ object PowerSaveModeHelper {
 
     // 处理 V1 机型 强制激活省电模式
     fun isPowerSaveForceActive(context: Context): Boolean {
-        if (SSDeviceUtils.isV1() || SSDeviceUtils.isV1Pro()) {
+        if (SSDeviceUtils.isV1Series() || SSDeviceUtils.isV1Pro()) {
             return true
         }